+++ packages/types/src/polls.ts
@@ -1,5 +1,14 @@
 export interface BasePolls {
   loadPolls(filter?: PollFilter, limit?: number): Promise<Poll[]>;
+  pollVote(pollId: string, optionId: number, anonymousId?: string): Promise<boolean>;
+  offPollReceived(): void;
+  onPollReceived(callback: (poll: Poll) => void): void;
+  offPollModified(): void;
+  onPollModified(callback: (poll: Poll) => void): void;
+  offPollDeleted(): void;
+  onPollDeleted(callback: (poll: Poll) => void): void;
+  offAllListeners(): void;
+  watchUserPollsReactions(userId: string): void;
 }
 
 export interface Poll {
@@ -16,6 +25,9 @@ export interface Poll {
   siteId: string;
   total: number;
   updatedAt: number;
+  expireAt: number;
+  currentUserVote: number;
+  changeType: string;
 }
 
 export enum PollFilter {