+++ packages/types/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.1.1 (2020-10-14)
+
+**Note:** Version bump only for package @arena-im/chat-types
+
+
+
+
+
+# 1.1.0 (2020-10-14)
+
+**Note:** Version bump only for package @arena-im/chat-types
+
+
+
+
+
 ## 1.0.1 (2020-09-23)